+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -7,7 +7,7 @@ Release History
 ++++++
 * `vm/vmss create --storage-sku`: can now specify the storage account sku for managed os and data disks separately.
 * `sig image-version`: Version names now consistently specified by  `--image-version -e`. `--image-version-name` deprecated.
-* `vm/vmss create --ephemeral-os-disk`: exposed parameter to create a vm with a local disk
+* `vm/vmss create --ephemeral-os-disk`: exposed parameter to create a vm/vmss with a local os disk.
 
 2.2.7
 ++++++