+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ===============
 
+2.2.8
+++++++
+* `vm/vmss create --storage-sku`: can now specify the storage account sku for managed os and data disks separately.
+* `sig image-version`: Version names now consistently specified by  `--image-version -e`. `--image-version-name` deprecated.
+* `vm/vmss create --ephemeral-os-disk`: exposed parameter to create a vm/vmss with a local os disk.
+* `snapshot create/update`: Added support for `--no-wait`.
+* `snapshot`: Added `wait` command.
+* `vm/vmss extension set --extension-instance-name`: can now specify the instance name of an extension.
+
 2.2.7
 ++++++
 * `image create`: expose storage-sku argument for setting the image's default storage account type